pyproject.toml
[tool.poetry]
name = "OCR_reader"
version = "0.1.0"
description = "Tool to extract text from a given PDF, JPEG, or PNG file"
authors = ["Joel Pires"]

[tool.poetry.dependencies]
python = "^3.4"
click = "^7.0"
tesserocr = "^2.3.0"
pdf2image = "^1.12.1"
yapf = "^0.30.0"
isort = "^4.3.18"
Pillow = "^7.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
